feat: abstract express server
