--- server/src/types.ts
@@ -1,5 +1,0 @@
-export interface User {
-  id: string;
-  email: string;
-  password_digest: string;
-}
